updated query variable names and created follow controller, model and routes
