+++ controllers/ProfileController.js
@@ -3,8 +3,10 @@ const profileQueries = require("../models/ProfileModel");
 
 const getUserPosts = async (req, res) => {
   try {
-    const { username } = req.query;
-    const query = await pool.query(profileQueries.getUserPosts, [username]);
+    const { visitedUsername } = req.query;
+    const query = await pool.query(profileQueries.getUserPosts, [
+      visitedUsername,
+    ]);
     res.send(query.rows);
   } catch (error) {
     console.log(error);
@@ -14,8 +16,10 @@ const getUserPosts = async (req, res) => {
 
 const getUserReplies = async (req, res) => {
   try {
-    const { username } = req.query;
-    const query = await pool.query(profileQueries.getUserReplies, [username]);
+    const { visitedUsername } = req.query;
+    const query = await pool.query(profileQueries.getUserReplies, [
+      visitedUsername,
+    ]);
     res.send(query.rows);
   } catch (error) {
     console.log(error);
@@ -25,8 +29,10 @@ const getUserReplies = async (req, res) => {
 
 const getUserLikes = async (req, res) => {
   try {
-    const { username } = req.query;
-    const query = await pool.query(profileQueries.getUserLikes, [username]);
+    const { visitedUsername } = req.query;
+    const query = await pool.query(profileQueries.getUserLikes, [
+      visitedUsername,
+    ]);
     res.send(query.rows);
   } catch (error) {
     console.log(error);
@@ -36,52 +42,10 @@ const getUserLikes = async (req, res) => {
 
 const getProfileContents = async (req, res) => {
   try {
-    const { userId, username } = req.query;
+    const { currentUserId, visitedUsername } = req.query;
     const query = await pool.query(profileQueries.getProfileContents, [
-      userId,
-      username,
-    ]);
-    res.send(query.rows[0]);
-  } catch (error) {
-    console.log(error);
-    res.status(500).send(error);
-  }
-};
-
-const getFollowStatus = async (req, res) => {
-  try {
-    const { userId, username } = req.query;
-    const query = await pool.query(profileQueries.getFollowStatus, [
-      userId,
-      username,
-    ]);
-    res.send(query.rows[0]);
-  } catch (error) {
-    console.log(error);
-    res.status(500).send(error);
-  }
-};
-
-const followUser = async (req, res) => {
-  try {
-    const { userId, username } = req.body;
-    const query = await pool.query(profileQueries.followUser, [
-      userId,
-      username,
-    ]);
-    res.send(query.rows[0]);
-  } catch (error) {
-    console.log(error);
-    res.status(500).send(error);
-  }
-};
-
-const unfollowUser = async (req, res) => {
-  try {
-    const { userId, username } = req.body;
-    const query = await pool.query(profileQueries.unfollowUser, [
-      userId,
-      username,
+      currentUserId,
+      visitedUsername,
     ]);
     res.send(query.rows[0]);
   } catch (error) {
@@ -95,7 +59,4 @@ module.exports = {
   getUserReplies,
   getUserLikes,
   getProfileContents,
-  getFollowStatus,
-  followUser,
-  unfollowUser,
 };
