+++ models/ProfileModel.js
@@ -135,45 +135,9 @@ FROM app_user AS a
 LEFT JOIN follow as f ON f."followerUserId" = $1 AND f."followedUserId" = a."userId"
 WHERE a."username" = $2`;
 
-const getFollowStatus = `SELECT 
-CASE 
-    WHEN EXISTS (
-        SELECT 1
-        FROM follow
-        WHERE "followerUserId" = $1
-          AND "followedUserId" = (SELECT "userId" FROM app_user WHERE "username" = $2)
-          AND "followStatus" = TRUE
-    ) THEN TRUE
-    ELSE FALSE
-END AS "followStatus";`;
-
-const followUser = `INSERT INTO follow ("followerUserId", "followedUserId", "followedDate", "followStatus")
-SELECT $1, "userId", CURRENT_DATE, TRUE
-FROM app_user
-WHERE "username" = $2
-ON CONFLICT ("followerUserId", "followedUserId")
-DO UPDATE SET "followStatus" = TRUE, "followedDate" = CURRENT_DATE
-RETURNING *;
-`;
-
-const unfollowUser = `WITH otherUser AS (
-SELECT "userId"
-FROM app_user
-WHERE "username" = $2
-LIMIT 1
-)
-UPDATE follow
-SET "followStatus" = FALSE
-WHERE "followerUserId" = $1
-AND "followedUserId" = (SELECT "userId" FROM otherUser)
-RETURNING *`;
-
 module.exports = {
   getUserPosts,
   getUserReplies,
   getUserLikes,
   getProfileContents,
-  getFollowStatus,
-  followUser,
-  unfollowUser,
 };
